frontend: Move ResourceCircularChart into its own common module
Under common/Resource/CircularChart.
